refactor: created the folder __components for each of the pages and refactored the agenda components, putting them all in the corresponding folder
