+++ src/ui/_components/atoms/index.ts
@@ -5,6 +5,5 @@ export { BigGoldInput } from './inputs/BigGoldInput'
 export { GoldSelectBox } from './inputs/GoldSelectBox'
 export { GoldButton } from './buttons/GoldButton'
 export { DefaultButton } from './buttons/DefaultButton'
-export { FooterButton } from './buttons/FooterButton'
 export type { GoldButtonProps } from './buttons/GoldButton'
 export type { InputProps, SelectBoxOptions, SelectBoxProps } from './inputs/InputProps'
