+++ src/ui/_components/molecules/index.ts
@@ -6,6 +6,3 @@ export { Header } from './titles/Header'
 export { SectionTitle } from './titles/SectionTitle'
 export { QuestionBox } from './cards/QuestionBox'
 export { TalkCard } from './cards/TalkCard'
-export { ModalHeader } from './modal/ModalHeader'
-export { ModalDescription } from './modal/ModalDescription'
-export { ModalFooter } from './modal/ModalFooter'
